Refactor build_sitefractions to parameter selection utils
