+++ espei/paramselect.py
@@ -24,8 +24,11 @@ from pycalphad import Database, Model, variables as v
 from sklearn.linear_model import LinearRegression
 
 from espei.core_utils import get_data, get_samples, canonicalize, canonical_sort_key, \
-    list_to_tuple, build_sitefractions
-from espei.parameter_selection.utils import endmembers_from_interaction, feature_transforms, shift_reference_state
+    list_to_tuple
+from espei.parameter_selection.utils import (
+    endmembers_from_interaction, feature_transforms, shift_reference_state,
+    build_sitefractions
+)
 from espei.parameter_selection.ternary_parameters import fit_ternary_formation_energy
 from espei.utils import PickleableTinyDB, sigfigs
 import espei.refdata
